feat: 完善lexer词法分析器功能
- 更新token.h头文件，改进词法单元定义
- 优化lexer.cpp核心词法分析逻辑
- 改进token.cpp词法单元实现
- 添加lexer调试测试文件
- 添加lexer简单功能测试文件

词法分析器现在支持更完整的token类型和更好的错误处理

+++ src/lexer/token.cpp
@@ -9,9 +9,9 @@ Token* Token::END_OF_FILE = new Token(lexer::END_OF_FILE);
 
 // 静态成员定义 - 直接初始化
 Type *Type::Int = new Type(NUM, "int", 4);
-Type *Type::Double = new Type(REAL, "double", 8);
+Type *Type::Double = new Type(DOUBLE, "double", 8);
 Type *Type::Char = new Type(CHAR, "char", 1);
-Type *Type::Float = new Type(REAL, "float", 8);
+Type *Type::Float = new Type(FLOAT, "float", 8);
 Type *Type::Bool = new Type(BOOL, "bool", 1);
 Type *Type::String = new Type(STR, "string", 0);
